chore: improve Portal exports
